Minor improvements +
better handling of wrong file/folder names, added function to delete all imgs from a folder (shorter code); now users can generate only the masks again even without having the 'generate_hybrid_frames' button selected

+++ scripts/deforum_helpers/vid2frames.py
@@ -10,7 +10,7 @@ def get_frame_name(path):
     name = os.path.splitext(name)[0]
     return name
 
-def vid2frames(video_path, video_in_frame_path, n=1, overwrite=True): 
+def vid2frames(video_path, video_in_frame_path, n=1, overwrite=True, only_get_fps=False): 
     #get the name of the video without the path and ext
     name = get_frame_name(video_path)
     if n < 1: n = 1 #HACK Gradio interface does not currently allow min/max in gr.Number(...) 
@@ -22,49 +22,52 @@ def vid2frames(video_path, video_in_frame_path, n=1, overwrite=True):
     else:
         if not os.path.exists(video_path):
             raise RuntimeError("Init video path or mask video path is not valid")
-
-    input_content = []
-    if os.path.exists(video_in_frame_path) :
-        input_content = os.listdir(video_in_frame_path)
-
-    # check if existing frame is the same video, if not we need to erase it and repopulate
-    if len(input_content) > 0:
-        #get the name of the existing frame
-        content_name = get_frame_name(input_content[0])
-        if not content_name.startswith(name):
-            overwrite = True
+            
     vidcap = cv2.VideoCapture(video_path)
     video_fps = vidcap.get(cv2.CAP_PROP_FPS)
 
-    # grab the frame count to check against existing directory len 
-    frame_count = int(vidcap.get(cv2.CAP_PROP_FRAME_COUNT)) 
-    
-    # raise error if the user wants to skip more frames than exist
-    if n >= frame_count : 
-        raise RuntimeError('Skipping more frames than input video contains. extract_nth_frames larger than input frames')
-    
-    expected_frame_count = math.ceil(frame_count / n) 
-    # Check to see if the frame count is matches the number of files in path
-    if overwrite or expected_frame_count != len(input_content):
-        shutil.rmtree(video_in_frame_path)
-        os.makedirs(video_in_frame_path, exist_ok=True) # just deleted the folder so we need to make it again
-        input_content = os.listdir(video_in_frame_path)
-    
-    if len(input_content) == 0:
-        success,image = vidcap.read()
-        count = 0
-        t=1
-        success = True
-        while success:
-            if state.interrupted:
-                return
-            if count % n == 0:
-                cv2.imwrite(video_in_frame_path + os.path.sep + name + f"{t:05}.jpg" , image)     # save frame as JPEG file
-                t += 1
+    if only_get_fps is False:
+        input_content = []
+        if os.path.exists(video_in_frame_path) :
+            input_content = os.listdir(video_in_frame_path)
+
+        # check if existing frame is the same video, if not we need to erase it and repopulate
+        if len(input_content) > 0:
+            #get the name of the existing frame
+            content_name = get_frame_name(input_content[0])
+            if not content_name.startswith(name):
+                overwrite = True
+        
+
+        # grab the frame count to check against existing directory len 
+        frame_count = int(vidcap.get(cv2.CAP_PROP_FRAME_COUNT)) 
+        
+        # raise error if the user wants to skip more frames than exist
+        if n >= frame_count : 
+            raise RuntimeError('Skipping more frames than input video contains. extract_nth_frames larger than input frames')
+        
+        expected_frame_count = math.ceil(frame_count / n) 
+        # Check to see if the frame count is matches the number of files in path
+        if overwrite or expected_frame_count != len(input_content):
+            shutil.rmtree(video_in_frame_path)
+            os.makedirs(video_in_frame_path, exist_ok=True) # just deleted the folder so we need to make it again
+            input_content = os.listdir(video_in_frame_path)
+        
+        if len(input_content) == 0:
             success,image = vidcap.read()
-            count += 1
-        print("Converted %d frames" % count)
-    else: print("Frames already unpacked")
+            count = 0
+            t=1
+            success = True
+            while success:
+                if state.interrupted:
+                    return
+                if count % n == 0:
+                    cv2.imwrite(video_in_frame_path + os.path.sep + name + f"{t:05}.jpg" , image)     # save frame as JPEG file
+                    t += 1
+                success,image = vidcap.read()
+                count += 1
+            print("Converted %d frames" % count)
+        else: print("Frames already unpacked")
     
     vidcap.release()
